+++ NEWS.md
@@ -1,5 +1,7 @@
 # glue (development version)
 
+* `glue_sql()` now always preserves the type of the column even in the presence of missing values (#130)
+
 # glue 1.4.0
 
 * `.envir = NULL` is now supported and is equivalent to passing `.envir = emptyenv()` (#140)